Move default img to public folder
