+++ dapps/kiosk/src/utils/constants.ts
@@ -5,3 +5,6 @@ export const TANSTACK_OWNED_KIOSK_KEY = 'TANSTACK_OWNED_KIOSK_KEY';
 export const TANSTACK_KIOSK_KEY = 'TANSTACK_KIOSK_KEY';
 export const TANSTACK_KIOSK_DATA_KEY = 'TANSTACK_KIOSK_DATA_KEY';
 export const TANSTACK_OWNED_OBJECTS_KEY = 'TANSTACK_OWNED_OBJECTS_KEY';
+
+// const public images
+export const DEFAULT_IMAGE = '/empty-image.svg';
